Solve Euler Problem #43
This solution is _really_ slowly and can probably be improved, but it
works for now -- just adds a lot of time to the runtime of the test
suite. Technically I think it takes ~100 seconds to run, which violates
Project Euler's 60-second cap, but I'll try to improve it in the future.