Address issues recommended by hlint
Almost all the suggestions made by hlint have been applied, with the
exception of one. On L231 of Poker.hs, hlint says to change

    reverse . sortBy (compare `on` length) . groupedCardValues

to

    sortBy (flip (compare `on` length)) . groupedCardValues

However, this change broke some code that uses the Euler.Poker module.
Theoretically it should work, so I'm not sure why it broke code, but it
did, so I'll have to investigate it a bit more before applying this
change.