Group data-related modules under Euler.Data
There seems to be no real consensus in the Haskell community about
module naming, vis-à-vis whether modules that are similar to stdlib
modules should be grouped at the top level or under a project name. In
other words, should it be Data.Bool.Euler or Euler.Data.Bool? Personally
I think the latter makes for nicer-looking documentation and
organization, so I'm going to standardize around that convention.