Upgrade to latest Stack resolver: lts-20.2
Previously we were unable to upgrade beyond lts-14.16 because the
"dates" package was not compatible with newer resolvers. However, I
forked the dates package and fixed that issue, so if we use that fork,
we can use the latest version of the Stack resolver, lts-20.2. We should
switch back to using the release version of dates as soon as possible,
but this solution will allow the project to be built and run on M1 Macs.