Mark Problem 25 as pending
This no longer runs in a timely manner. I'm not sure why. I honestly
haven't run this code in years; not only is today the first time I have
tried running it in ages, but it's the first time I've tried running it
on Apple Silicon (an M2 Mac Studio), so maybe Haskell running on Apple
Silicon is an issue, but regardless, this test no longer works.